fruit-ripeness/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
onnxruntime==1.19.2
pillow==10.4.0
numpy==1.26.4
